requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.1.2
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
